lossless-claw-5b6: extract engine transcript readers
